+++ ouroboros-consensus-cardano-tools/src/Cardano/Api/KeysShelley.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE DataKinds                  #-}
 {-# LANGUAGE DeriveAnyClass             #-}
 {-# LANGUAGE DerivingVia                #-}
 {-# LANGUAGE FlexibleInstances          #-}
@@ -7,11 +8,6 @@
 {-# LANGUAGE OverloadedStrings          #-}
 {-# LANGUAGE ScopedTypeVariables        #-}
 {-# LANGUAGE TypeFamilies               #-}
-
--- The Shelley ledger uses promoted data kinds which we have to use, but we do
--- not export any from this API. We also use them unticked as nature intended.
-{-# LANGUAGE DataKinds                  #-}
-{-# OPTIONS_GHC -Wno-unticked-promoted-constructors #-}
 
 -- DUPLICATE -- adapted from: cardano-api/src/Cardano/Api/KeysShelley.hs