+++ ouroboros-consensus-protocol/CHANGELOG.md
@@ -1,5 +1,15 @@
-# CHANGELOG
+# Changelog for ouroboros-consensus-protocol
 
-This package is part of a bundle of packages. Please consult `CHANGELOG.md` in
-the package `ouroboros-consensus` at the same version as this package to see the
-current changelog that applies also to this package.
+# Changelog entries
+
+<a id='changelog-0.5.0.0'></a>
+## 0.5.0.0 - 2023-04-14
+
+### Breaking
+
+- Extract `ouroboros-consensus-protocol` from the bundle of packages.
+
+## Before 0.5.0.0
+
+Before this version, `ouroboros-consensus-protocol` lived in a bundle of
+packages with `ouroboros-consensus`, thus the changelog was the same.